New function added.
* source/inc/global.hxx (toVector): new function added.

* source/inc/numeric/lpmodel.hxx
* source/numeric/lpmodel.cxx: changed argument types of several variable
boundary related functions from unsigned long to size_t.

* source/inc/numeric/type.hxx: added javadoc comments to 2 enums.

* source/numeric/lpsimplex.cxx: removed unused header include.  Did a
major rewrite of the initial solution search algorithm for bounded
revised simplex method (BoundedRevisedSimplexImpl).  Extracted iteration
variable initialization part into its own method to make it easier to
follow.

+++ scratch/scsolver/source/inc/numeric/lpmodel.hxx
@@ -64,9 +64,9 @@ public:
 	void setCostVector( const std::vector<double>& );
 	void deleteCostVectorElements( const std::vector<size_t> & );
 
-	double   getVarBound( unsigned long, Bound ) const;
-	void     setVarBound( unsigned long, Bound, double );
-	bool isVarBounded( unsigned long, Bound ) const;
+	double   getVarBound( size_t, Bound ) const;
+	void     setVarBound( size_t, Bound, double );
+	bool isVarBounded( size_t, Bound ) const;
 
 	Goal getGoal() const;
 	void setGoal( Goal );
@@ -89,7 +89,7 @@ public:
 	double getRhsValue( size_t ) const;
 	void setRhsValue( size_t, double );
 	std::vector<Equality> getEqualityVector() const;
-	Equality getEqualityByRowId( unsigned long ) const;
+	Equality getEqualityByRowId( size_t ) const;
 	void addConstraint( const std::vector<double>&, Equality, double );
 	void setStandardConstraintMatrix( const ::scsolver::numeric::Matrix&, const ::scsolver::numeric::Matrix& );
 	void deleteConstraintMatrixColumns( const std::vector<size_t>& );
